make pixels have doubles

+++ include/Pixel.hh
@@ -8,16 +8,16 @@ class Pixel
 public:
     Pixel();
     Pixel(int x, int y);
-    Pixel(int x, int y, int a, int r, int g, int b);
+    Pixel(int x, int y, double a, double r, double g, double b);
     ~Pixel();
     Pixel operator+(const Pixel& p) const;
     void clamp();
     int x_;
     int y_;
-    int a_;
-    int r_;
-    int g_;
-    int b_;
+    double a_;
+    double r_;
+    double g_;
+    double b_;
 
     std::string describe() const;
 };
